switched values to int

+++ src/pathfinder/configuration/Configuration.java
@@ -6,9 +6,9 @@ public class Configuration {
 	public final static float WALLDISTANCE = 30;	
 	public final static float INITIAL_WALLDISTANCE = 100;
 	
-	public final static float OBSTACLE_SIZE = 40;		
-	public final static float OBSTACLE_OFFSET = 5;
-	public final static float TOTAL_OBSTACLE_SIZE = OBSTACLE_OFFSET + OBSTACLE_SIZE;
+	public final static int OBSTACLE_SIZE = 40;		
+	public final static int OBSTACLE_OFFSET = 5;
+	public final static int TOTAL_OBSTACLE_SIZE = OBSTACLE_OFFSET + OBSTACLE_SIZE;
 	
 	public final static Direction INITIAL_DIRECTION = Direction.RIGHT;
